Refactor MenuEvents to extract MenuSearch logic
